requirements.txt
google-analytics-data
pandas
azure-storage-blob